Requested test failing CI; mv'ed to manual integration file; code clean up
