+++ client/dom/test/genesearch.integration.spec-manual.ts
@@ -1,13 +1,41 @@
 import tape from 'tape'
 import { string2variant } from '../genesearch.ts'
 import { hg38 } from '../../test/testdata/genomes'
+import * as d3s from 'd3-selection'
+import { addGeneSearchbox } from '../genesearch.ts'
+import { Menu } from '../menu'
+import { detectOne } from '../../test/test.helpers.js'
 
 /* Tests
     - string2variant() - HGVS deletion and delins variants
+	- Gene search results on keyup
 
 *** Tests cannot be run on CI because the fasta file is not available ***
 Run manually as needed. 
 */
+
+/**************
+ helper functions
+***************/
+
+function getHolder() {
+	return d3s.select('body').append('div').style('padding', '5px').style('margin', '5px')
+}
+
+function getRow(holder) {
+	return holder.append('div').style('border', '1px solid #aaa').style('padding', '5px')
+}
+
+function getSearchBox(holder, opts = {}) {
+	const _opts = {
+		genome: hg38,
+		tip: new Menu({ padding: '' }),
+		row: getRow(holder)
+	}
+
+	const args = Object.assign(_opts, opts)
+	addGeneSearchbox(args)
+}
 
 /**************
  test sections
@@ -38,7 +66,7 @@ tape('string2variant() - HGVS deletion and delins variants', async test => {
 	expected = {
 		isVariant: true,
 		chr: 'chr2',
-		ref: 'AGCTG',//on CI this shows as undefined because the fasta file is not available
+		ref: 'AGCTG', //on CI this shows as undefined because the fasta file is not available
 		alt: '-'
 	}
 	test.deepEqual(
@@ -73,7 +101,7 @@ tape('string2variant() - HGVS deletion and delins variants', async test => {
 		isVariant: true,
 		chr: 'chr2',
 		pos: 119955155,
-		ref: 'AGCTG',//on CI this shows as undefined because the fasta file is not available
+		ref: 'AGCTG', //on CI this shows as undefined because the fasta file is not available
 		alt: 'TTTTT'
 	}
 	test.deepEqual(variant, expected, 'Should parse HGVS string into a Delins variant object')
@@ -86,5 +114,35 @@ tape('string2variant() - HGVS deletion and delins variants', async test => {
 	expected = undefined
 	test.equal(variant, expected, 'Should return undefined for invalid position for delins format')
 
+	test.end()
+})
+
+tape('Gene search results on keyup', async test => {
+	test.timeoutAfter(10000)
+	const holder = getHolder()
+	const tip = new Menu({ padding: '' })
+	getSearchBox(holder, { tip })
+	const searchInput = holder.select('input').node() as HTMLInputElement
+	const gene = 'p53'
+
+	//Check if the search box is functional
+	/** Leave this
+	 * Options populate slightly slower than tgit he menu
+	 * Need to wait for options to appear before testing
+	 */
+	const result = await detectOne({
+		selector: '.sja_menuoption',
+		target: tip.dnode,
+		trigger() {
+			searchInput.value = gene
+			searchInput.dispatchEvent(new KeyboardEvent('keyup'))
+		}
+	})
+	test.true(result.textContent.includes(gene.toUpperCase()), 'Should display matching gene options')
+
+	if (test['_ok']) {
+		if (tip.dnode) tip.dnode.remove()
+		holder.remove()
+	}
 	test.end()
 })
