+++ client/dom/test/genesearch.unit.spec.ts
@@ -53,26 +53,6 @@ tape('Default gene search box', async test => {
 	test.ok(searchInput.tagName == 'INPUT', 'Should create an input element')
 	test.equal(searchInput.placeholder, 'Gene, position, dbSNP', 'Should display the default placeholder text')
 
-	// //Check if the search box is functional
-	// await detectOne({
-	// 	target: tip.dnode,
-	// 	trigger() {
-	// 		searchInput.value = 'BRCA1'
-	// 		searchInput.dispatchEvent(new Event('keyup'))
-	// 	}
-	// })
-	// /** Leave this
-	//  * Options populate slightly slower than the menu
-	//  * Need to wait for options to appear before testing
-	//  */
-	// const options = await detectGte({
-	// 	selector: '.sja_menuoption',
-	// 	target: tip.dnode,
-	// 	count: 2
-	// })
-	// const foundGenes = options.filter((d: HTMLElement) => d.textContent?.includes('BRCA1'))
-	// test.equal(options.length, foundGenes.length, 'Should display matching gene options')
-
 	if (test['_ok']) {
 		if (tip.dnode) tip.dnode.remove()
 		holder.remove()
